{"schema":"decodex/commit/1","summary":"Modularize selection flow affordances","authority":"manual"}
